requirements.txt
imagehash
opencv-python
pillow
numpy
ConfigParser